Reverting entire framework
Reverting back to Nginx with RTMP module, as this now seems to work without building Nginx specifically with that module (as it seems to have required in the past). The NodeJS variant was having stability issues and not accepting streams properly for an unclear reason. Will likely attempt to reimplement that variant at some point if I can debug the issue.